Porting over to new Bro plugin API.
This goes with the plugins-2.3 branch now.

The log/dns*compiler tests still fail because of type-casting problem
in base/protocols/dns/main.bro:171:

    error, type of parameter 2 does not match function, have
    ref<vector<ref<DNS::Info>>> but expected
    ref<vector<ref<LibBro::BroAny>>>

I had previously worked around that by changing the Bro script.

+++ bro/src/Pac2Analyzer.h
@@ -56,7 +56,7 @@ public:
 	void Init() override;
 	void Done() override;
 	void DeliverStream(int len, const u_char* data, bool orig) override;
-	void Undelivered(int seq, int len, bool orig) override;
+	void Undelivered(uint64 seq, int len, bool orig) override;
 	void EndOfData(bool is_orig) override;
 	void FlipRoles() override;
 
@@ -84,8 +84,8 @@ public:
 	void Init() override;
 	void Done() override;
 	void DeliverPacket(int len, const u_char* data, bool orig,
-			   int seq, const IP_Hdr* ip, int caplen) override;
-	void Undelivered(int seq, int len, bool orig) override;
+			   uint64 seq, const IP_Hdr* ip, int caplen) override;
+	void Undelivered(uint64 seq, int len, bool orig) override;
 	void EndOfData(bool is_orig) override;
 	void FlipRoles() override;
